# api-gateway/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Learning Platform API
  description: API для управления образовательной платформой
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /user/{userId}:
    get:
      tags:
        - User Info Controller
      summary: Получить информацию о другом пользователе
      description: Позволяет получить пользовательскую информацию о другом пользователе
      operationId: getOtherUserInfo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /user/me:
    get:
      tags:
        - User Info Controller
      summary: Получить свою информацию
      description: Позволяет получить пользовательскую информацию от пользователя
      operationId: getMyUserInfo
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
...
  /slot/my:
    get:
      tags:
        - Time Slot Controller
      summary: Выдать информацию обо всех слотах ментора
      description: Выдать информацию о слотах и об участниках. Требуются права MENTOR
      operationId: getMentorSlotsInfo
      responses:
        '200':
          description: Выдан список слотов
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSlotInfoDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSlotInfoDto'
...
  /module/{courseId}/{moduleId}:
    get:
      tags:
        - Module Controller
      summary: Получить модуль по ID
      description: Возвращает информацию о модуле по его идентификатору
      operationId: getModuleById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о модуле
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
    delete:
      tags:
        - Module Controller
      summary: Удалить модуль
      description: Позволяет удалить модуль. Требуются права ADMIN или MENTOR
      operationId: deleteModule
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Модуль успешно удален
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
  /course/{courseId}:
    get:
      tags:
        - Course Controller
      summary: Получить курс по ID
      description: Возвращает информацию о курсе по его идентификатору
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о курсе
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
    delete:
      tags:
        - Course Controller
      summary: Удалить курс
      description: Позволяет удалить курс. Требуются права ADMIN или MENTOR
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Курс успешно удален
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                type: object
  /course/all:
    get:
      tags:
        - Course Controller
      summary: Получить все курсы
      description: Возвращает список всех курсов, включая неактивные
      operationId: getAllCourses
      responses:
        '200':
          description: Список всех курсов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
...
  /admin/user/{userId}:
    get:
      tags:
        - Admin User Info Controller
      summary: Получить информацию о другом профиле.
      description: Позволяет получить информацию о профиле другого пользователя.
      operationId: adminGetOtherUserInfo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /admin/user/me:
    get:
      tags:
        - Admin User Info Controller
      summary: Получить информацию о своем профиле.
      description: Позволяет получить информацию о своем профиле.
      operationId: adminGetMyUserInfo
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /admin/slot/all:
    get:
      tags:
        - Admin Mentor Time Slot Controller
      summary: Получить информацию обо всех слотах ментора
      description: Получить информацию о слотах и участниках. Требуются права администратора.
      operationId: adminGetMentorSlotsInfo
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Выдан список слотов
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageMentorSlotInfoDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageMentorSlotInfoDto'
...
  /admin/course/{courseId}:
    get:
      tags:
        - Admin Course Controller
      summary: Получить курс по ID
      description: Возвращает информацию о курсе по его идентификатору
      operationId: adminGetCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о курсе
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
  /admin/course/all:
    get:
      tags:
        - Admin Course Controller
      summary: Получить все курсы
      description: Возвращает список всех курсов, включая неактивные
      operationId: adminGetAllCourses
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Список всех курсов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCourseDto'
components:
  schemas:
    MentorTimeSlotDto:
      type: object
      properties:
        id:
          type: integer
          description: ID слота
          format: int64
          example: 1
        mentorId:
          type: integer
          description: ID ментора
          format: int64
          example: 1
        rqUId:
          type: string
          description: UUID запроса
          example: 6e8f4e02-c91c-465f-b22d-7f102fca381b
        startTime:
          type: string
          description: Начало открытого слота
          format: date-time
        endTime:
          type: string
          description: Окончание открытого слота
          format: date-time
        slotType:
          type: string
          description: Тип слота
          example: INDIVIDUAL
          enum:
            - INDIVIDUAL
            - GROUP
        slotMeetingType:
          type: string
          description: Тип планируемой встречи
          example: COMMUNICATION
          enum:
            - ACQUAINTANCE
            - COMMUNICATION
            - ACCEPTING
        maxParticipants:
          type: integer
          description: Максимальное число участников
          format: int32
          example: 10
        meetingLink:
          type: string
          description: Ссылка на встречу
          example: https://www.meet.ru/<uuid>
        description:
          type: string
          description: Описание слота
          example: Знакомство и обсуждение плана дальнейшего взаимодействия
        createdAt:
          type: string
          description: Дата и время создания создания слота
          format: date-time
        isActive:
          type: boolean
          description: Признак активности слота
          example: true
      description: Информация о слоте
...
    ModuleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        moduleTitle:
          type: string
        moduleOrderNumber:
          type: integer
          format: int32
        moduleContent:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CourseDto:
      type: object
      properties:
        id:
          type: integer
          description: ID курса
          format: int64
          example: 1
        courseTitle:
          type: string
          description: Название курса
          example: Основы Spring Boot
        courseDescription:
          type: string
          description: Описание курса
          example: Курс по основам Spring Framework
        isActive:
          type: boolean
          description: Флаг активности курса
          example: true
        createdAt:
          type: string
          description: Дата создания курса
          format: date-time
        author:
          $ref: '#/components/schemas/UserInfoDto'
        modules:
          type: array
          description: Список модулей курса
          items:
            $ref: '#/components/schemas/ModuleDto'
      description: Информация о курсе
    UserInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
            - MENTOR
        firstName:
          type: string
        lastName:
          type: string
        tgNickname:
          type: string
        tgChatId:
          type: integer
          format: int64
      description: ID автора курса
      example: 5
...
    MentorSlotInfoDto:
      type: object
      properties:
        slotDto:
          $ref: '#/components/schemas/MentorTimeSlotDto'
        participants:
          type: array
          description: Информация об участниках слота
          items:
            $ref: '#/components/schemas/UserInfoDto'
      description: Подробная информация о слоте ментора и об участниках в нем
...
    PageMentorSlotInfoDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/MentorSlotInfoDto'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    PageCourseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean